+++ app-frontend/src/app/components/projects/projectItem/projectItem.module.js
@@ -1,3 +1,4 @@
+/* global BUILDCONFIG */
 import angular from 'angular';
 import { get } from 'lodash';
 
@@ -81,7 +82,11 @@ class ProjectItemController {
         const layer = L.tileLayer(
             this.projectService.getProjectTileURL(this.item, {
                 token: this.authService.token()
-            })
+            }),
+            {
+                maxNativeZoom: BUILDCONFIG.TILES_MAX_ZOOM,
+                maxZoom: BUILDCONFIG.VISUAL_MAX_ZOOM
+            }
         );
 
         this.getMap().then(m => {