+++ app-frontend/src/app/pages/projects/edit/edit.module.js
@@ -142,7 +142,8 @@ class ProjectsEditController {
             {token: this.authService.token()}
         );
         let layer = L.tileLayer(url, {
-            maxZoom: 30
+            maxNativeZoom: BUILDCONFIG.TILES_MAX_ZOOM,
+            maxZoom: BUILDCONFIG.VISUAL_MAX_ZOOM
         });
 
         this.getMap().then(m => {